# gpt-sovits-backend/GPT-SoVITS.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GPT-SoVITS",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/admin/users/{user_id}/status": {
      "put": {
        "summary": "更新用户状态（激活/禁用）",
        "deprecated": false,
        "description": "更新用户状态（激活/禁用）",
        "operationId": "put_admin_update_user_status",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "路径参数: user_id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "is_active": {
                    "type": "string",
                    "description": "JSON 字段: is_active"
                  }
                },
                "required": [
                  "is_active"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/user/delete-account": {
      "delete": {
        "summary": "删除用户账户",
        "deprecated": false,
        "description": "删除用户账户",
        "operationId": "delete_user_delete_account",
        "tags": [
          "user"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/profile": {
      "get": {
        "summary": "获取用户个人资料",
        "deprecated": false,
        "description": "获取用户个人资料",
        "operationId": "get_user_get_profile",
        "tags": [
          "user"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "summary": "更新用户个人资料",
        "deprecated": false,
        "description": "更新用户个人资料",
        "operationId": "put_user_update_profile",
        "tags": [
          "user"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "avatar_url": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "avatar_url"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "profile": {
                          "type": "object",
                          "properties": {
                            "avatar_url": {
                              "type": "string"
                            },
                            "created_at": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            },
                            "id": {
                              "type": "string"
                            },
                            "is_active": {
                              "type": "boolean"
                            },
                            "is_verified": {
                              "type": "boolean"
                            },
                            "last_login_at": {
                              "type": "string"
                            },
                            "role": {
                              "type": "integer"
                            },
                            "username": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "avatar_url",
                            "created_at",
                            "email",
                            "id",
                            "is_active",
                            "is_verified",
                            "last_login_at",
                            "role",
                            "username"
                          ]
                        }
                      },
                      "required": [
                        "profile"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "message",
                    "success"
                  ]
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/statistics": {
      "get": {
        "summary": "获取用户统计信息",
        "deprecated": false,
        "description": "获取用户统计信息",
        "operationId": "get_user_get_user_statistics",
        "tags": [
          "user"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/tasks/history": {
      "get": {
        "summary": "获取用户任务历史",
        "deprecated": false,
        "description": "获取用户任务历史",
        "operationId": "get_user_get_task_history",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "查询参数: page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "查询参数: per_page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "查询参数: type",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "查询参数: status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/uploads": {
      "get": {
        "summary": "获取用户上传文件列表",
        "deprecated": false,
        "description": "获取用户上传文件列表",
        "operationId": "get_user_get_user_uploads",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "查询参数: page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "查询参数: per_page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "查询参数: type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/uploads/{upload_id}": {
      "delete": {
        "summary": "删除上传文件",
        "deprecated": false,
        "description": "删除上传文件",
        "operationId": "delete_user_delete_upload",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "description": "路径参数: upload_id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "examples": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}